ООП 2.3 Дз 1. Задание 1 выполнено.
